Refactor: - SpelExpressionParser extracted from SpelExpression - Flavour extracted from SpelExpressionParser + renamed to SpelFlavour - Typer renamed to SpelTyper - SpelExpression renamed to CompiledSpelExpression
